Refactoring: move modules into sub dirs.
